backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.3.0=py310hca03da5_0
aniso8601=9.0.1=pyhd3eb1b0_0
anyio=3.5.0=py310hca03da5_0
appnope=0.1.2=py310hca03da5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h1a28f6b_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310hca03da5_0
blas=1.0=openblas
bleach=4.1.0=pyhd3eb1b0_0
brotlipy=0.7.0=py310h1a28f6b_1002
bzip2=1.0.8=h620ffc9_4
c-ares=1.18.1=h1a28f6b_0
ca-certificates=2022.07.19=hca03da5_0
cairo=1.16.0=h29d4eff_2
certifi=2022.9.24=py310hca03da5_0
cffi=1.15.1=py310h22df2f2_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
chex=0.1.5=pypi_0
click=8.0.4=py310hca03da5_0
colorama=0.4.5=pypi_0
commonmark=0.9.1=pypi_0
contourpy=1.0.5=pypi_0
cryptography=37.0.1=py310h834c97f_0
cycler=0.11.0=pypi_0
dalle-mini=0.1.1=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py310hc377ac9_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dm-tree=0.1.7=pypi_0
docker-pycreds=0.4.0=pypi_0
eigen=3.4.0=hc021e02_0
einops=0.4.1=pypi_0
emoji=2.1.0=pypi_0
entrypoints=0.4=py310hca03da5_0
etils=0.8.0=pyhd8ed1ab_0
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.9=hb7217d7_0
ffmpeg=4.2.2=h04105a8_0
fftw=3.3.9=h1a28f6b_1
filelock=3.8.0=pypi_0
flask=2.0.3=pyhd3eb1b0_0
flask-cors=3.0.10=pypi_0
flask-restful=0.3.9=pyhd8ed1ab_0
flax=0.6.0=pypi_0
fontconfig=2.14.0=hfb34624_0
fonttools=4.37.4=pypi_0
freetype=2.11.0=h1192e45_0
ftfy=6.1.1=pypi_0
gettext=0.21.0=h826f4ad_0
giflib=5.2.1=h1a28f6b_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
glib=2.69.1=h98b2900_1
gmp=6.2.1=hc377ac9_3
gnutls=3.6.15=h887c41c_0
graphite2=1.3.14=hc377ac9_1
grpc-cpp=1.47.1=h503f348_6
gst-plugins-base=1.14.1=hf0a386a_0
gstreamer=1.14.1=he09cfb7_0
harfbuzz=4.3.0=hb1b0ec1_0
hdf5=1.12.1=nompi_hf9525e8_104
huggingface-hub=0.10.0=pypi_0
icu=68.1=hc377ac9_0
idna=3.4=py310hca03da5_0
importlib_resources=5.2.0=pyhd3eb1b0_1
ipykernel=6.15.2=py310hca03da5_0
ipython=8.4.0=py310hca03da5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
itsdangerous=2.0.1=pyhd3eb1b0_0
jax=0.3.20=pyhd8ed1ab_0
jaxlib=0.3.20=cpu_py310h18a4bca_0
jedi=0.18.1=py310hca03da5_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.2.0=pypi_0
jpeg=9e=h1a28f6b_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.16.0=py310hca03da5_0
jupyter=1.0.0=py310hca03da5_8
jupyter_client=7.3.5=py310hca03da5_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.11.1=py310hca03da5_0
jupyter_server=1.18.1=py310hca03da5_0
jupyterlab=3.4.4=py310hca03da5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.15.2=py310hca03da5_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.4=pypi_0
krb5=1.19.2=h3b8d789_0
lame=3.100=h1a28f6b_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libabseil=20220623.0=cxx17_h28b99d4_4
libblas=3.9.0=15_osxarm64_openblas
libcblas=3.9.0=15_osxarm64_openblas
libclang=12.0.0=default_hc321e17_4
libcurl=7.84.0=hc6d1d07_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.8=h1a28f6b_5
libedit=3.1.20210910=h1a28f6b_0
libev=4.33=h642e427_1
libffi=3.4.2=hc377ac9_4
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libiconv=1.16=h1a28f6b_2
libidn2=2.3.1=h1a28f6b_0
liblapack=3.9.0=15_osxarm64_openblas
libllvm12=12.0.0=h12f7ac0_4
libnghttp2=1.47.0=h232270b_1
libopenblas=0.3.21=h269037a_0
libopus=1.3=h1a28f6b_1
libpng=1.6.37=hb8d0fd4_0
libpq=12.9=h65cfe13_3
libprotobuf=3.21.6=hb5ab8b9_1
libsodium=1.0.18=h1a28f6b_0
libsqlite=3.39.3=h76d750c_0
libssh2=1.10.0=hb80f160_3
libtasn1=4.16.0=h1a28f6b_0
libtiff=4.4.0=had003b8_0
libunistring=0.9.10=h1a28f6b_0
libvpx=1.10.0=hc377ac9_0
libwebp=1.2.4=h68602c7_0
libwebp-base=1.2.4=h1a28f6b_0
libxml2=2.9.14=h8c5e841_0
libxslt=1.1.35=h9833966_0
libzlib=1.2.12=h03a7124_3
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.3=hc377ac9_0
markupsafe=2.1.1=py310h1a28f6b_0
marshmallow=3.18.0=pypi_0
matplotlib=3.6.0=pypi_0
matplotlib-inline=0.1.6=py310hca03da5_0
mistune=0.8.4=py310h1a28f6b_1000
msgpack=1.0.4=pypi_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py310hca03da5_0
nbconvert=6.4.4=py310hca03da5_0
nbformat=5.5.0=py310hca03da5_0
ncurses=6.3=h1a28f6b_3
nest-asyncio=1.5.5=py310hca03da5_0
nettle=3.7.3=h84b5d62_1
nomkl=3.0=0
notebook=6.4.12=py310hca03da5_0
nspr=4.33=hc377ac9_0
nss=3.74=h142855e_0
numpy=1.23.3=pypi_0
numpy-base=1.23.1=py310h742c864_0
openblas=0.3.21=hca03da5_0
openblas-devel=0.3.21=hca03da5_0
openh264=1.8.0=h98b2900_0
openssl=1.1.1q=h1a28f6b_0
opt_einsum=3.3.0=pyhd8ed1ab_1
optax=0.1.3=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
pcre=8.45=hc377ac9_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310h4d1bdd5_1
pip=21.2.4=py310hca03da5_0
pixman=0.40.0=h27ca646_0
ply=3.11=py310hca03da5_0
prometheus_client=0.14.1=py310hca03da5_0
promise=2.3=pypi_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
protobuf=3.20.3=pypi_0
psutil=5.9.2=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.13.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310hca03da5_0
pyqt=5.15.7=py310hc377ac9_0
pyqt5-sip=12.11.0=pypi_0
pyrsistent=0.18.0=py310h1a28f6b_0
pysocks=1.7.1=py310hca03da5_0
python=3.10.6=hbce4517_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py310hca03da5_0
python_abi=3.10=2_cp310
pytorch=1.12.1=py3.10_0
pytz=2022.1=py310hca03da5_0
pyyaml=6.0=pypi_0
pyzmq=23.2.0=py310hc377ac9_0
qt-main=5.15.2=ha2d02b5_7
qt-webengine=5.15.9=h2903aaf_4
qtconsole=5.3.2=py310hca03da5_0
qtpy=2.2.0=py310hca03da5_0
qtwebkit=5.212=h0f11f3c_4
re2=2022.06.01=h9a09cb3_0
readline=8.1.2=h1a28f6b_1
regex=2022.9.13=pypi_0
requests=2.28.1=py310hca03da5_0
rich=11.2.0=pypi_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.1=pypi_0
send2trash=1.8.0=pyhd3eb1b0_1
sentry-sdk=1.9.9=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=63.4.1=py310hca03da5_0
shortuuid=1.0.9=pypi_0
sip=6.6.2=py310hc377ac9_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sniffio=1.2.0=py310hca03da5_1
soupsieve=2.3.2.post1=py310hca03da5_0
sqlite=3.39.3=h1058600_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.13.1=py310hca03da5_0
testpath=0.6.0=py310hca03da5_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.12.1=pypi_0
toml=0.10.2=pyhd3eb1b0_0
toolz=0.12.0=pypi_0
torchaudio=0.12.1=py310_cpu
torchvision=0.13.1=py310_cpu
tornado=6.2=py310h1a28f6b_0
tqdm=4.64.1=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.22.2=pypi_0
typing-extensions=4.3.0=py310hca03da5_0
typing_extensions=4.3.0=py310hca03da5_0
tzdata=2022e=h04d1e81_0
unidecode=1.3.6=pypi_0
urllib3=1.26.12=py310hca03da5_0
vqgan-jax=0.0.1=pypi_0
wandb=0.13.3=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310hca03da5_1
websocket-client=0.58.0=py310hca03da5_4
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py310hca03da5_0
x264=1!152.20180806=h1a28f6b_0
xz=5.2.6=h1a28f6b_0
zeromq=4.3.4=hc377ac9_0
zipp=3.8.0=py310hca03da5_0
zlib=1.2.12=h5a0b063_3
zstd=1.5.2=h8574219_0
